requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
backports.asyncio.runner==1.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
exceptiongroup==1.3.0
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
